Extract firmware-job persistence into persistence.py
Pulls _load_jobs / _persist_jobs / _prune_history out of
controllers/firmware/controller.py into a new persistence.py
submodule as free functions taking the controller as first arg
(the canonical devices/firmware_sync.py pattern). The controller
keeps thin one-line delegate methods so existing test patches on
`controller._load_jobs` / `controller._persist_jobs` /
`controller._prune_history` (~10 sites under tests/controllers/firmware/)
keep working unchanged.

_ACTIVE_JOB_STATUSES moves from controller.py to constants.py
alongside the other history-pool constants — avoids a circular
import since persistence needs it and constants is a leaf module.

controller.py shrinks 2083 → 1978 lines (-105). First PR of the
broader firmware/controller.py split-then-clean cycle (the file
is still grandfathered over the 800-line cap; further concerns
to extract: lifecycle/cancel, subprocess composition, queue/dispatch).
